backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy==1.26.0
python-multipart==0.0.6
pydantic-settings==2.1.0